* Protocol 8.6x support (otb remains to be updated!). * Updated doc files. * Patched autostacking from otserv. * Misc. fixes.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@3931 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ luascript.h
@@ -335,7 +335,7 @@ class LuaScriptInterface
 
 		static int32_t getField(lua_State* L, const char* key);
 		static uint32_t getFieldU32(lua_State* L, const char* key);
-		static void setField(lua_State* L, const char* index, int32_t val);
+		static void setField(lua_State* L, const char* index, double val);
 		static void setField(lua_State* L, const char* index, const std::string& val);
 		static std::string getFieldString(lua_State* L, const char* key);
 		static void setFieldBool(lua_State* L, const char* index, bool val);
